requirements.txt
PyExecJS==1.5.1
requests==2.24.0
beautifulsoup4==4.9.3
